healthcare/ich_segmentation/requirements.txt
nibabel
scikit_learn
pillow
matplotlib
google-cloud-storage
cassandra-driver
pandas
seaborn